+++ src/store/src/Bridge/Pinecone/Store.php
@@ -13,6 +13,7 @@ namespace Symfony\AI\Store\Bridge\Pinecone;
 
 use Probots\Pinecone\Client;
 use Probots\Pinecone\Resources\Data\VectorResource;
+use Symfony\AI\Platform\Vector\NullVector;
 use Symfony\AI\Platform\Vector\Vector;
 use Symfony\AI\Store\Document\Metadata;
 use Symfony\AI\Store\Document\VectorDocument;
@@ -135,9 +136,17 @@ final class Store implements ManagedStoreInterface, StoreInterface
         );
 
         foreach ($result->json()['matches'] as $match) {
+            $vector = new Vector($match['values']);
+
+            if (!($options['include_vectors'] ?? true)) {
+                unset($match['values']);
+
+                $vector = new NullVector();
+            }
+
             yield new VectorDocument(
                 id: $match['id'],
-                vector: new Vector($match['values']),
+                vector: $vector,
                 metadata: new Metadata($match['metadata']),
                 score: $match['score'],
             );